backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
psycopg2-binary
django-cors-headers
dj-rest-auth
django-allauth>=0.66.0
requests
cryptography
PyJWT